services/api-gateway/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
aiokafka==0.8.0
kafka-python==2.0.2
python-jose[cryptography]==3.3.0
redis==5.0.1
prometheus-client==0.20.0
